cleanup

+++ app/Http/Controllers/OAuth/OAuthProviderController.php
@@ -6,20 +6,20 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\OAuthProviderResource;
 use App\Integrations\OAuth\OAuthProviderService;
 use App\Models\OAuthProvider;
-use App\Models\Workspace;
 use Illuminate\Support\Facades\Auth;
 
 class OAuthProviderController extends Controller
 {
-    // TODO user_id
     public function index()
     {
-        $providers = Workspace::first()->providers()->get();
+        /** @var \App\Models\User $user */
+        $user = Auth::user();
+
+        $providers = $user->oauthProviders()->get();
 
         return OAuthProviderResource::collection($providers);
     }
 
-    // TODO authorization
     public function connect(OAuthProviderService $service)
     {
         return response()->json([
@@ -29,16 +29,13 @@ class OAuthProviderController extends Controller
 
     public function handleRedirect(OAuthProviderService $service)
     {
-        /** @var \App\Models\User $user */
-        $user = Auth::user();
-
         $driverUser = $service->getDriver()->getUser();
 
         $provider = OAuthProvider::query()
             ->updateOrCreate(
                 [
-                    'workspace_id' => $user->workspaces()->first()->id,
-                    'provider' => 'google',
+                    'user_id' => Auth::id(),
+                    'provider' => $service,
                     'provider_user_id' => $driverUser->getId(),
                 ],
                 [
@@ -52,11 +49,12 @@ class OAuthProviderController extends Controller
         return OAuthProviderResource::make($provider);
     }
 
-    // TODO authorization
     public function destroy(OAuthProvider $provider)
     {
+        $this->authorize('delete', $provider);
+
         $provider->delete();
 
-        return response()->json([]);
+        return response()->json();
     }
 }
